TIP-860: Move Attribute related api tests
