+++ tests/back/Pim/Structure/EndToEnd/Attribute/ExternalApi/CreateAttributeEndToEnd.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Attribute;
+namespace AkeneoTest\Pim\Structure\EndToEnd\Attribute\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
@@ -9,7 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
 /**
  * @group ce
  */
-class CreateAttributeIntegration extends ApiTestCase
+class CreateAttributeEndToEnd extends ApiTestCase
 {
     public function testHttpHeadersInResponseWhenAnAttributeIsCreated()
     {
